광고 조회 통신 연결 미완성
