+++ app/build.gradle
@@ -2,8 +2,6 @@ plugins {
     id 'com.android.application'
 }
 
-
-
 android {
     compileSdkVersion 33
     buildToolsVersion "30.0.3"
